+++ shaku/src/container/container.rs
@@ -1,10 +1,9 @@
 use crate::container::ComponentMap;
-use crate::module::{HasComponent, Module};
+use crate::module::Module;
 use crate::provider::ProviderFn;
-use crate::ContainerBuilder;
-use crate::Interface;
-use crate::Provider;
+use crate::{ContainerBuilder, HasComponent};
 use crate::{HasProvider, ProvidedInterface};
+use crate::{HasSubmodule, Interface};
 use std::error::Error;
 use std::sync::Arc;
 
@@ -14,19 +13,65 @@ use std::sync::Arc;
 /// [`Module`]: trait.Module.html
 /// [`ContainerBuilder`]: struct.ContainerBuilder.html
 /// [`Container::default`]: struct.Container.html#method.default
-pub struct Container<M: Module> {
-    pub(crate) module: M,
-    pub(crate) provider_overrides: ComponentMap,
+pub struct Container<'m, M: Module> {
+    pub(crate) inner: ContainerData<'m, M>,
 }
 
-impl<M: Module> Default for Container<M> {
+/// Internal container variants and associated data
+pub(crate) enum ContainerData<'m, M: Module> {
+    /// The root container. It owns the module and provider overrides.
+    Root {
+        module: M,
+        provider_overrides: ComponentMap,
+    },
+    /// A submodule of the root module. It references the submodule instance
+    /// owned by the root module and the provider overrides. A submodule
+    /// container cannot mutate its components.
+    Submodule {
+        module: &'m M,
+        provider_overrides: &'m ComponentMap,
+    },
+}
+
+impl<'m, M: Module> ContainerData<'m, M> {
+    /// Get a reference to the module
+    fn module_ref(&self) -> &M {
+        match self {
+            ContainerData::Root { module, .. } => module,
+            ContainerData::Submodule { module, .. } => module,
+        }
+    }
+
+    /// Get a mutable reference to the module. Only returns `Some` for root
+    /// modules.
+    fn module_mut(&mut self) -> Option<&mut M> {
+        match self {
+            ContainerData::Root { module, .. } => Some(module),
+            ContainerData::Submodule { .. } => None,
+        }
+    }
+
+    /// Get a reference to the provider overrides
+    fn provider_overrides(&self) -> &ComponentMap {
+        match self {
+            ContainerData::Root {
+                provider_overrides, ..
+            } => provider_overrides,
+            ContainerData::Submodule {
+                provider_overrides, ..
+            } => provider_overrides,
+        }
+    }
+}
+
+impl<M: Module> Default for Container<'static, M> {
     /// Build a default container. Same as `ContainerBuilder::new().build()`.
     fn default() -> Self {
         ContainerBuilder::new().build()
     }
 }
 
-impl<M: Module> Container<M> {
+impl<'m, M: Module> Container<'m, M> {
     /// Get a reference to the component registered with the interface `I`. The ownership of
     /// the component is shared via `Arc`.
     ///
@@ -57,7 +102,7 @@ impl<M: Module> Container<M> {
     where
         M: HasComponent<I>,
     {
-        Arc::clone(self.module.get_ref())
+        Arc::clone(self.inner.module_ref().get_ref())
     }
 
     /// Create a service using the provider registered with the interface `I`.
@@ -96,10 +141,11 @@ impl<M: Module> Container<M> {
     where
         M: HasProvider<I>,
     {
-        self.provider_overrides
+        self.inner
+            .provider_overrides()
             .get::<ProviderFn<M, I>>()
             .map(|provider_fn| provider_fn(self))
-            .unwrap_or_else(|| M::Impl::provide(self))
+            .unwrap_or_else(|| M::provide(self))
     }
 
     /// Get a reference to the component registered with the interface `I`.
@@ -131,14 +177,15 @@ impl<M: Module> Container<M> {
     where
         M: HasComponent<I>,
     {
-        Arc::as_ref(self.module.get_ref())
+        Arc::as_ref(self.inner.module_ref().get_ref())
     }
 
     /// Get a mutable reference to the component registered with the interface `I`.
     ///
     /// If the component is jointly owned (an `Arc<I>` reference to the component exists outside of
     /// this container), then `None` will be returned as it is unsafe to take a
-    /// mutable reference without exclusive ownership of the component.
+    /// mutable reference without exclusive ownership of the component. `None`
+    /// will also be returned if this is a submodule container.
     ///
     /// # Example
     /// ```
@@ -167,6 +214,21 @@ impl<M: Module> Container<M> {
     where
         M: HasComponent<I>,
     {
-        Arc::get_mut(self.module.get_mut())
+        Arc::get_mut(self.inner.module_mut()?.get_mut())
+    }
+
+    /// Create a submodule container. The created container references this
+    /// container and can be used to resolve components/providers exposed by the
+    /// submodule.
+    pub fn sub_container<N: Module>(&self) -> Container<'_, N>
+    where
+        M: HasSubmodule<N>,
+    {
+        Container {
+            inner: ContainerData::Submodule {
+                module: self.inner.module_ref().get_submodule(),
+                provider_overrides: self.inner.provider_overrides(),
+            },
+        }
     }
 }