+++ shaku/src/container/mod.rs
@@ -2,12 +2,12 @@
 
 #[allow(clippy::module_inception)]
 mod container;
-mod container_build_context;
 mod container_builder;
+mod module_build_context;
 
 pub use self::container::Container;
-pub use self::container_build_context::ContainerBuildContext;
 pub use self::container_builder::ContainerBuilder;
+pub use self::module_build_context::ModuleBuildContext;
 
 #[cfg(not(feature = "thread_safe"))]
 type AnyType = dyn anymap::any::Any;