+++ shaku/src/module.rs
@@ -1,9 +1,14 @@
-use crate::{Component, ContainerBuildContext, Interface, ProvidedInterface, Provider};
+use crate::{Container, Interface, ModuleBuildContext, ProvidedInterface};
+use std::error::Error;
 use std::sync::Arc;
 
 /// A module represents a group of services. By implementing traits such as
 /// [`HasComponent`] on a module, service dependencies are checked at compile
 /// time. At runtime, modules hold the components they are associated with.
+///
+/// Modules can also use other modules as submodules, importing specific
+/// services for the root module's use. For more details, see the [`module`]
+/// macro.
 ///
 /// Modules are usually created via the [`module`] macro.
 ///
@@ -32,13 +37,21 @@ use std::sync::Arc;
 pub trait Module: Sized + 'static {
     /// Create the module instance by resolving the components this module
     /// provides.
-    fn build(context: &mut ContainerBuildContext<Self>) -> Self;
+    fn build(context: &mut ModuleBuildContext<Self>) -> Self;
+}
+
+/// Indicates that a module contains a submodule.
+pub trait HasSubmodule<M: Module> {
+    fn get_submodule(&self) -> &M;
 }
 
 /// Indicates that a module contains a component which implements the interface.
 pub trait HasComponent<I: Interface + ?Sized>: Module {
-    /// The concrete component which implements the interface
-    type Impl: Component<Self, Interface = I>;
+    /// Resolve the component during module build. Usually this involves calling
+    /// [`ModuleBuildContext::resolve`] with the implementation.
+    ///
+    /// [`ModuleBuildContext::resolve`]: struct.ModuleBuildContext.html#method.resolve
+    fn resolve(context: &mut ModuleBuildContext<Self>) -> Arc<I>;
 
     /// Get a reference to the stored component. This is used when resolving the
     /// component.
@@ -51,8 +64,11 @@ pub trait HasComponent<I: Interface + ?Sized>: Module {
 
 /// Indicates that a module contains a provider which implements the interface.
 pub trait HasProvider<I: ProvidedInterface + ?Sized>: Module {
-    /// The concrete provider which implements the interface
-    type Impl: Provider<Self, Interface = I>;
+    /// Build the service. Usually this involves calling the implementation's
+    /// [`Provider::provide`] method.
+    ///
+    /// [`Provider::provide`]: trait.Provider.html#tymethod.provide
+    fn provide(container: &Container<Self>) -> Result<Box<I>, Box<dyn Error + 'static>>;
 }
 
 /// Create a [`Module`] which is associated with some components and providers.
@@ -63,8 +79,8 @@ pub trait HasProvider<I: ProvidedInterface + ?Sized>: Module {
 /// `MySecondModule` in the example below.
 ///
 /// ## Circular dependencies
-/// Note that this macro will detect circular dependencies at compile time. The
-/// error that is thrown will be something like
+/// This macro will detect circular dependencies at compile time. The error that
+/// is thrown will be something like
 /// "overflow evaluating the requirement `Component2: shaku::component::Component<TestModule>`".
 ///
 /// It is still possible to compile with a circular dependency if the module is
@@ -91,11 +107,13 @@ pub trait HasProvider<I: ProvidedInterface + ?Sized>: Module {
 /// }
 ///
 /// // MySecondModule implements HasComponent<dyn MyComponent> by using
-/// // MyModule's implementation.
+/// // MyModule's implementation. It also implements Module and
+/// // HasSubmodule<MyModule>.
 /// module! {
 ///     MySecondModule {
 ///         components = [],
 ///         providers = [],
+///
 ///         use MyModule {
 ///             components = [MyComponent],
 ///             providers = []
@@ -131,29 +149,35 @@ macro_rules! module {
                 // It would be nice to prefix the property with something like
                 // "__di_", but macro_rules does not support concatenating
                 // idents on stable.
-                $component: ::std::sync::Arc<<$component as $crate::Component<Self>>::Interface>
-            ),*
+                $component: ::std::sync::Arc<<$component as $crate::Component<Self>>::Interface>,
+            )*
             $($(
-                $submodule: $submodule
-            ),*)?
+                $submodule: $submodule,
+            )*)?
         }
 
         impl $crate::Module for $module {
-            fn build(context: &mut $crate::ContainerBuildContext<Self>) -> Self {
+            fn build(context: &mut $crate::ModuleBuildContext<Self>) -> Self {
                 Self {
                 $(
-                    $component: context.resolve::<<$component as $crate::Component<Self>>::Interface>()
-                ),*
+                    $component: <Self as $crate::HasComponent<
+                        <$component as $crate::Component<Self>>::Interface
+                    >>::resolve(context),
+                )*
                 $($(
-                    $submodule: context.build_submodule::<$submodule>()
-                ),*)?
+                    $submodule: context.as_submodule($submodule::build),
+                )*)?
                 }
             }
         }
 
         $(
         impl $crate::HasComponent<<$component as $crate::Component<Self>>::Interface> for $module {
-            type Impl = $component;
+            fn resolve(
+                context: &mut $crate::ModuleBuildContext<Self>
+            ) -> ::std::sync::Arc<<$component as $crate::Component<Self>>::Interface> {
+                context.resolve::<$component>()
+            }
 
             fn get_ref(&self) -> &::std::sync::Arc<<$component as $crate::Component<Self>>::Interface> {
                 &self.$component
@@ -167,13 +191,30 @@ macro_rules! module {
 
         $(
         impl $crate::HasProvider<<$provider as $crate::Provider<Self>>::Interface> for $module {
-            type Impl = $provider;
+            fn provide(container: &$crate::Container<Self>) -> ::std::result::Result<
+                ::std::boxed::Box<<$provider as $crate::Provider<Self>>::Interface>,
+                ::std::boxed::Box<dyn ::std::error::Error + 'static>
+            > {
+                $provider::provide(container)
+            }
         }
         )*
 
+        $($(
+        impl $crate::HasSubmodule<$submodule> for $module {
+            fn get_submodule(&self) -> &$submodule {
+                &self.$submodule
+            }
+        }
+        )*)?
+
         $($($(
         impl $crate::HasComponent<$sub_component> for $module {
-            type Impl = <$submodule as $crate::HasComponent<$sub_component>>::Impl;
+            fn resolve(
+                context: &mut $crate::ModuleBuildContext<Self>
+            ) -> ::std::sync::Arc<$sub_component> {
+                context.as_submodule($submodule::resolve)
+            }
 
             fn get_ref(&self) -> &::std::sync::Arc<$sub_component> {
                 self.$submodule.get_ref()
@@ -186,9 +227,14 @@ macro_rules! module {
         )*)*)?
 
         $($($(
-        impl $crate::HasProvider<$sub_provider> for $module
-        {
-            type Impl = <$submodule as $crate::HasProvider<$sub_provider>>::Impl;
+        impl $crate::HasProvider<$sub_provider> for $module {
+            fn provide(container: &$crate::Container<Self>) -> ::std::result::Result<
+                ::std::boxed::Box<$sub_provider>,
+                ::std::boxed::Box<dyn ::std::error::Error + 'static>
+            > {
+                let sub_container = container.sub_container::<$submodule>();
+                <$submodule as $crate::HasProvider<$sub_provider>>::provide(&sub_container)
+            }
         }
         )*)*)?
     };